Fix for thirddegree/HatchitGame#28.
This should resolve the issue where the two raptors with
MeshRendererComponents were not being displayed in Release. We were not
verifying the iterator returned by nlohmann::json::find before accessing
the object it contained in Release. As a result of this, if the requested
JSON property did not exist an exception was thrown by nlohmann::json.